Update `ICPSource` to not store ICPs as doubles.
This is another PR towards not using doubles for ICP values. `ICPSource`
represented the balance as string `balance` along with a helper method
`icpBalance` that parses this string and returns it as a double.

This PR replaces the `balance` string and `icpBalance` method with an
`icpBalance` of type ICP. In follow-up PRs further uses of doubles for
ICP will be removed.