Store memos for old transactions and recalculate transaction types (#85)
We only started storing memo values for transactions from May 19th.

This means we are not able to detect 'Stake Neuron' transactions from before this time and so in turn are unable to detect 'Top-up Neuron' transactions. This means neurons created before this time must be refreshed manually.

This PR first sets all of the missing memo values and then recalculates every transaction type. Once this is done we will know the address of every neuron created via the NNS Dapp and so will be able to auto refresh all of their balances.