NU-124 Give users a button to increase the stake of a neuron (#116)
Currently to increase the stake of a neuron, users need to find the 'Stake Neuron' transaction and send additional ICP to the same address. This is not very intuitive or user friendly. This commit adds a button onto each neuron which allows users to easily increase the stake of their selected neuron.