Script to deploy from CI artefact (#364)
# Motivation
People wish to deploy to a testnet for reasons including, but not limited to:
- Not being able to build locally, due to Mac M1 chips
- Not wanting to waste time building when CI has already built a wasm file.

# Changes
- Added a script to deploy a wasm file downloaded from CI.

Note: github assets cannot be downloaded by anonymous users.  In the envisaged use case, a human user will be choosing a build artifact so they might as well click "Download" while they are about it.

Note: It is assumed that the download directory is ~/Downloads.  This is the default on Mac and Linux.  We can support other use cases as the need arises.