feat: infinite scroll component (#408)
# Motivation

Proposals and other list are not fetched integrally when displayed but through subsets that are displayed when the user scrolls a specified distance from the bottom or top of the page.

This PR introduces an "infinite scroll" component which implements such behavior for any lists ("for each").

# Changes

- new `<InfiniteScroll />` component

# Notes

The component relies on the [Intersection Observer API](https://developer.mozilla.org/en-US/docs/Web/API/Intersection_Observer_API).

Such API is [not supported](https://caniuse.com/?search=intersection%20Observer) on Safari version older than v12.1 on desktop and v12.2 on mobile. There browsers are out of the scope of the browser we are currently supporting ("latest 2"). That's why the PR does not include any polyfill or some sort of custom code for old browsers.

# Tests

Unit test a basic since there is no implementation of the Intersection Observer in jest. It will need e2e tests for a proper behavioral tests.