Make a note of the sha sum (#479)
# Motivation
It is useful to be able to verify which code is deployed.  The subnet dashboard provides a trustworthy sha256sum of the wasm.  We need a mechanism of linking that sha256sum back to one or more commits.  There can be several matching commits if there are code changes that do not affect the wasm build.

# Changes
Add a git note with the wasm sha.

Notes can be accessed with:
```
$ git fetch origin refs/notes/*:refs/notes/*
$ git notes --ref=testnet/wasm-sha
$ git notes --ref=mainnet/wasm-sha
```