Let the svelte linter fail (#522)
# Motivation
The svelte linter rules have been or will soon be updated.  This causes, as expected, a lot of errors.  According to our plan, we are going to allow these errors to appear for a while, to give people time to fix them.  However for that github must not block PR merges on account of lint failures.

# Changes
* Put the svelte lint job in a separate task.
* Make a new dummy task that controls whether we may merge, rather than listing requirements in the admin panel.