Redirect sometimes (#518)
# Motivation
We plan to deploy the voting tab first.  For this we will need to be able to switch between svelte and flutter, depending on which tab the user is on.

# Changes

REDIRECT_TO_LEGACY was a binary value, true or false.  We need a third value that lets us see tab changes before they are deployed.  We now have:

Existing string values: "true" "1" "false" "0"
New string values:  "prod", "staging"