feat: web workers and sign-out on invalid auth delegation (#582)
# Motivation

This PR introduces the support for web workers to the nns-dapp svelte app and a feature that checks periodically in the background if the authentication delegation is still valid. In case of invalidity, the cronjob initiate an automatic sign-out of the user. 

# Changes

Worker:

- rollup dependency [@surma/rollup-plugin-off-main-thread](https://www.npmjs.com/package/@surma/rollup-plugin-off-main-thread) to use the worker as module
- enhance `rollup.config.js` to bundle both an app `bundle.js` and a new `worker.js`

Feature:

- `worker.ts` as main entry of the worker which also listen for messages of the client side
- `auth.worker.ts` to check periodically the auth status
- sync auth and worker in `App.svelte`
- new utils to retrieve delegation and identity saved by agent-js in the local storage
- add `worker.services.ts` which initialize the worker on the client side including the message listener

# Context

Feature wise, this PR has the same goal as PR [#567 - Logout expired session](https://github.com/dfinity/nns-dapp/pull/567). We shall decide which approach we rather like.