Tests may use any version of chrome (#613)
# Motivation
The e2e tests should work with the latest few versions of chrome.  The major version of chrome changes every 2 weeks.  The package.json should be stable because it encodes our expectations of what we want to test.  The lockfile is there to fix the specific choice that matches our expectations.  Therefore `package.json` should have something like `latest` or maybe `*`.  The lockfile should be bumped with a newer chromedriver every 2 weeks.