feat: new transaction store, context and init amount step (#657)
# Motivation

To create new transaction, user have to select and enter several data such as selecting the source account, selecting the destination address or account and the amount of ICP.

This PR add a store that contains these information and scope contextually its usage to the wizard (to avoid props going up and down and to scope logic withing components - i.e. steps).

In addition, to display the result of the selection, this PR also initialize the third step of the "New transaction" wizard. 

# Changes

- new `transaction.store`
- new `NewTransactionAmount` step - component
- extract common function that return the constant transaction fee as string
- insert `NewTransactionAmount` into wizard
- use new store in a wizard `context`
- use context in `NewTransactionDestination` (renamed) and `NewTransactionSource` (renamed) to select source and destination information
- bind input and selection of the account in `NewTransactionDestination`