feat: re-filter proposals returned by governance canister (#697)
# Motivation

The governance canister interprets queries with empty filter (e.g. topics=[]) has "any" queries and returns proposals anyway. On the contrary, the Flutter app displays nothing if one filter is empty.

The Flutter app does not simply display nothing when a filter is empty but re-filter the results provided by the backend.

This PR aims to reproduce the same behavior as Flutter app.

# Note

We could have implemented a feature that just empty the proposals store if no filters is provided - i.e. no backend query if filter is empty but we cannot be 100% sure it ultimately matches the exact same behavior as Flutter in production.

In addition, since we also implemented certified queries, this can also lead to some glitch if users changes the filters selection often really quickly.

Speaking of, if we stick to update calls, we might someday implement a feature that disable the filters selection until the certified queries are finished. Like "Please wait to update your filters until data are certified". For the future.

# Edge case glitch

An edge case glitch might happen when user search filter X, update calls take long, user modify filter Y meanwhile and returned call is re-processed with filter Y until ultimately the second update calls kicks. Not wrong but user might experience a UX that does "display results -> nothing found -> displayed results". This PR however try to minimize this by updating the "Nothing found" info only if the store is certified.

# Changes

- re-process the returned proposals of the governance canister - i.e. checks if content matches currently selected filters
- add `certified` info to store
- process "nothing found" info only if store is certified