Testnet environments (#816)
# Motivation
We need two testnet environments:
- One with prod-like settings, currently deployed as https://nns-prod.winning.gold
- One with svelte and one with flutter.  Both are provided by https://nns-both.winning.gold

Historically bot have been deployed as the "testnet" network, but with different copies of the nns-dapp repo, to keep the two canister_ids.json separate.

Now we are able to add additional network entries in dfx, so we can deploy both testnet deployments from a single repo.  This simplifies deployment.

# Changes
Define nnsdapp-prod and nnsdapp-both in dfx.json.