Disable guess if --open is specified (#884)
# Motivation
If individual steps are specified in deploy.sh, guessing what is required should be disabled.  This is already true for most steps.

However `./deploy.sh --open` should just open the nns-dapp in a browser.  But because the `GUESS=false` that is set for all other similar flags is omitted for `--open`, this tries to deploy as well.

# Changes
- Set `GUESS=false` for `--open` in `deploy.sh`.