feat: navbar style and stots (#1010)
# Motivation

Because we introduced the menu and removed tabs, the page name isn't displayed on views anymore. Therefore this PR remove the static "NETWORK NERVOUS SYSTEM" and replaces it with a dynamic title.

In addition, navbar used to be different on main screens as on details page. This PR aligns both style - i.e. one header / navbar style and behavior for the all app.

Finally, the "logout" option was not available on the detail page, so this PR makes the option available too.

# Changes

- make the title in `Header` dynamic
- refactor `HeadlessLayout` to use common `Header`
- refactor `Toolbar` to use it in `footer` (as existing) and newly in `header` too