Install idl2json (#1101)
# Motivation
When sns canisters are created via the sns-wasms canister, not dfx, they are no longer added to the `canister_ids.json`.  It is possible to use `idl2json` and `jq` to use the `sns` cli response to populate `canister_ids.json`, and so create  
a better developer experience.  This requires `idl2json` to be available.

# Changes
* Add code to install idl2json to the setup script