Fix: Do not fail spawning neurons (#1125)
# Motivation

Governance canister will be deployed first with the new spawning functionality. Until the new nns-dapp is release, the spawning neurons break the UI and show a blank page.

Fix the error and show only 0 ICP and that's it. At this point, we can't use the new state nor the new property. We just need to avoid the blank page.

# Changes

* stateInfo can be undefined for unknown states.
* Do not show state information if stateInfo is undefined.

# Tests

* Type changes only