Include identities for use against the SNS testnet (#1116)
# Motivation

There is a deployed SNS with some predefined roles attributed to principals that belong to the included identities. To allow for others at dfinity to test the SNS I've cut this PR to make use of these identities accessible. I've also included a README.md that lists what has been included and gives a brief overview of how to use them.

# Changes

- Include identities and some other files for use with an SNS