feat: proposal open until duration expires (#1129)
# Motivation

A proposal should remain open for vote according its duration not its status because it could be accepted or declined when the majority is reached but still accept votes.

# Logic

To detect if a proposal is still open for voting, the field `deadline_timestamp_seconds` should be observed. Even though it is always set by the backend it is (for some unknown reason) set as optional in the definition. Therefore a fallback should be implemented. This PR implement such a fallback as the existing code of the dashboard.

# Changes

- new function `isProposalOpenForVote` that uses `deadlineTimestampSeconds` or fallback to some logic implemented as in the dashboard
- replace checks on status `OPEN` in list (filter "Hide Open proposal ....") and display of voting card
- bump nns-js