Make cycles fast (#1138)
# Motivation
Cycle minting rate limiting is now applied in the CMC canister.  Creating an SMS is expensive enough that we run into the rate limit quite quickly (although it seems to me that we hit the limit much too quickly - need to check).  If we hit the rate limiter, no more cycles can be minted for an hour, which significantly slows down test deployments, especially if there are bugs and the deployment needs to be repeated.

# Changes
* Reduce the rate limit time from 1 hour to 2 minutes.  Long enough that we can see that the limit exists but not so long that testnet deployment times are impacted seriously.
* Deleted a CMC patch that we don't need any more.