refactor: load sns summaries and swaps with derived store (#1179)
# Motivation

As discussed with @lmuntaner this PR refactor the way we load the sns summaries and swaps in store.

Instead of getting the information summaries and swaps from the backend, cleaning and concatening these (match swap and summary with root id + cleaning optional candid types + sort) and applying to the stores, this pr get the information from backend, apply these to a global store and make the information available to the app through a derived store.

Think a bit the "redux selector way" rather than "preparing data and putting these in store afterwards".

# Note

I did modify the global sns/projects store but I did not modify the context store used in the detail - i.e. the context store will still get data that are prepared in advance. Open to discussion but since the context store for the detail page contains both summary, swap and commitment, it would have need to split it in two or three stores. Feels like an over complication that does not bring that much clarity in the architecture.

# Changes

- introduces a new `sns.store` that contains sns answers from backend - i.e. sns summary, swaps but also proposals
- refactor `project.store` to derive above store
- move stores from `project` to `sns` if more "sns" or abstraction "projects" related