L2-956: Ignore last commitment error on notify (#1265)
# Motivation

User of the last commitment received an error notification because the notify was failing. The problem is that the backend might be faster to notify than the frontend. That causes the project to change state to Committed and the subsequent notify to fail.

Catch that error specifically and ignore it.

# Changes

* Check for last commitment notify error in the participateInSwap sns service.
* Export project util commitmentExceedsAmountLeft to be used in participateInSwap.

# Tests

* Test exported project util commitmentExceedsAmountLeft
* Test the service for the new use case.