feat: dompurify as chunk instead of lazy (#1283)
# Motivation

Instead of lazy loading dompurify as a js assets, this PR adds it as a chunked library. We are now using the sanitizer in more places, the probability that the user will have to download it is bigger, therefore the explicit lazy loading isn't that necessary anymore and we can rely on chunks. In addition, it makes the code imperative which can be easier to introduce the component `<HTML />` we briefly commented in one of my last PR.

# Changes

- replace dompurify js assets with dependency
- adapt import and sanitizer function accordingly
- update jest to make dompurify supported for nodejs tests