fix: proposals infinite scroll for new design (#1287)
# Motivation

The new grid design is not compatible with the existing infinite scrolling approach. In addition, hiding cards with CSS to achieve the feature wasn't that ideal from an accessibility perspective.

# Notes

- the infinite scroll observer has to be modified to observe the last element of the "list" instead of observing an element that is displayed after x%
- to prevent issue, we add a new option to disable the infinite scroll when there is no more data to fetch
- the order of the proposals that were added to the store were not maintained between query and update results which  lead to some ui shift that had for end effect to trigger the infinite scrolling which triggered new searches with incorrect proposal id 

e.g.

query results -> store: id 10, 9, 8, 7, 6
trigger, next search index: 6
update results -> store: 8, 7, 6, 10, 9
repaint, trigger, next search index: 9
query results  -> store: : 8, 7, 6, 10, 9, 5, 4, 3, 2, 1
update results -> store: messy store

I could have fixed the issue but, for the proposal store, the order is actually not relevant and fixing it would not prevent future issue. Therefore I created a derived store that secure the order.

# Future development

Even though it is solve, as currently, the ui can become a bit shaky depending of the ui rerender. I ask my self if we should not, when update calls result matches query calls, not re-render the ui and this even though we use keyed Svelte each. Not for now but, for the future.

# PR

- [X] need `gix-component` PR https://github.com/dfinity/gix-components/pull/45

# Changes

- new `sortedProposals` derived store
- remove cards from DOM instead of hiding it to fulfill specific client side proposal's filter too
- style voting filters container
- display spinner when more proposals need to be loaded and skeleton when proposals are fetched for the first time
- disable infinite scroll when nothing else to fetch
- refactor `replaceAndConcatenateProposals` to filter existing proposals