Kill earlier runs (#1342)
# Motivation

If a developer (e.g. myself) pushes a branch, then notices a change that needs to be made and pushes again, the earlier job still runs to completion even though it is no longer of interest.  It would be nice to kill the earlier job when it is superseded.

Github's concurrency provides setting to do just this:

https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-only-cancel-in-progress-jobs-or-runs-for-the-current-workflow

I have opted for the configuration where each workflow has its own groups, so that a kill command from one workflow doesn't kill others.  This lets us choose for which workflows this policy is applied.  I have applied the concurrency limit to every workflow but we may decide that it is only appropriate for some.

# Changes
* Set each workflow to kill earlier runs of the same workflow on the same branch.