FIX: Cache sns wrapper by identity (#1547)
# Motivation

We are using different identities to interact with SNSes. For now logged
in and anonymous.

Therefore, we need to cache the sns wrapper by identity so that we don't
use the wrong wrapper. For example, using the anonymous identity wrapper
to get the balance.

# Changes

* Change implementation of "wrappers" to cache by identity principal
instead of only by certified or not.
* Map transfer errors and show a proper message according to each error.

# Tests

* No changes. Same functionality.