Fix lm voting power (#1574)
# Motivation

Voting power calculation doesn't match the actual value of voting power.

Calculation is used in modal to set or increase dissolve delay.

The issue was that it was missing the staked maturity.

It also includes a big refactor which I hope makes it clear how the
voting power is calculated.

# Changes

* Rename votingPower neuron util to neuronVotingPower and change the
params. It now expects the neuron and a new dissolve delay. The new
dissolve delay is useful to show the future voting power in the increase
dissolve delay modal.
* neuronVotingPower uses a new function votingPower. Which expects all
the params and data necessary to calculate voting power. It users the
new "bonusMultiplier" function.
* dissolveDelayMultiplier and ageMultiplier use a new "bonusMultiplier"
function.
* new bonusMultiplier helper function.
* Moved functions around to put helpers next to each other for better
context.
* Remove the voting power UI in SNS modals.

# Tests

* Add some scenarios and fix tests with new api.